% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgrm.R
\name{annotate_power}
\alias{annotate_power}
\title{Add power annotations to a result set that's been merged with PGRM}
\usage{
annotate_power(annotated_results, LOUD = FALSE)
}
\arguments{
\item{annotated_results}{A data.table of results annotated by PGRM}

\item{LOUD}{If TRUE then progress info is printed to the terminal. Default FALSE}
}
\value{
A data.table of the results with a column Power added which includes 80\% power calculations (alpha=0.05)
}
\description{
This function adds a power calculation to a result set that has been annotated with PGRM
}
