% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgrm.R
\name{annotate_results}
\alias{annotate_results}
\title{Annotate a result set with the PGRM}
\usage{
annotate_results(
  results,
  use_allele_dir = T,
  ancestry = "all",
  build = "hg19",
  phecode_version = "V1.2",
  calculate_power = FALSE,
  annotate_CI_overlap = T,
  LOUD = TRUE
)
}
\arguments{
\item{results}{A data frame with results of a test cohort; columns for SNP, phecode, cases, controls, odds_ratio, P (see demo files like results_BBJ for example)}

\item{use_allele_dir}{If TRUE, direction of effect is used when assessing if an association is replicated}

\item{ancestry}{A string that specifies ancestry of the PGRM that is then used to annotate the results file. Options EAS, EUR, AFR, SAS, AMR, ALL. Default ALL}

\item{build}{A string indicating the genome reference build used in the results table. Options hg19, hg37. Default is hg19.}

\item{phecode_version}{A string indicating the phecode version used in the results table. Currently only V1.2 is supported, which is the default}

\item{calculate_power}{If TRUE then power calculations will be conducted using case and control counts from the results file. Necessary for get_AE(). Default FALSE}

\item{annotate_CI_overlap}{If TRUE then a column called "annotate_CI_overlap" is added to the table, values:
(\strong{overlap}: 95\% CIs of PGRM and test cohort overlap, \strong{test_cohort_greater}: 95\% CI of test cohort greater than PGRM, \strong{PGRM_greater}: 95\% CI of PGRM greater than test cohort)}

\item{LOUD}{If TRUE then progress info is printed to the terminal. Default FALSE}
}
\value{
A data.table of the results file annotated with columsn from the PGRM
}
\description{
This function annotates a result from a test cohort with information from the PGRM
}
