% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map.R
\docType{data}
\name{exclude_ranges}
\alias{exclude_ranges}
\title{Exclude ranges for phecode controls}
\format{
A data table with the following columns:
\itemize{
\item \code{phecode}: Character vector of phecode
\item \code{range_start}: Numeric vector of the start of the exclude range
\item \code{range_end}: Numeric vector of the end of the exclude range
}
}
\source{
\url{https://phewascatalog.org}
}
\usage{
exclude_ranges
}
\description{
The data are artificial and do not correspond to real patients.
}
\seealso{
\code{\link[=get_pheno]{get_pheno()}}
}
\keyword{datasets}
