% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgrm.R
\name{get_AER}
\alias{get_AER}
\title{Calculate the acual:expected ratio (A:E) of a test cohort}
\usage{
get_AER(annotated_results, LOUD = TRUE)
}
\arguments{
\item{annotated_results}{A data table of results that have been annotated with the PGRM}

\item{LOUD}{If TRUE then progress info is printed to the terminal. Default TRUE}
}
\value{
An numeric value of the actual:expected ratio of the result set
}
\description{
This function calculates the actual:expected a test cohort that has been annotated with PGRM.
}
