% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgrm.R
\name{get_powered_rate}
\alias{get_powered_rate}
\title{Calculate the percent of associations that are powered in a test set}
\usage{
get_powered_rate(annotated_results, LOUD = TRUE)
}
\arguments{
\item{annotated_results}{A data table of results that have been annotated with the PGRM}

\item{LOUD}{If TRUE then progress info is printed to the terminal. Default TRUE}
}
\value{
An numeric value of the percent of associations that are powered
}
\description{
This function calculates \% of powered associations for a test cohort that has been annotated with PGRM.
}
