% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgrm.R
\name{get_RR_power}
\alias{get_RR_power}
\title{Replicaiton rate (RR) of powered associations}
\usage{
get_RR_power(results, LOUD = TRUE)
}
\arguments{
\item{results}{An data.table of results, annotated with the pgrm}

\item{LOUD}{If TRUE then progress info is printed to the terminal. Default TRUE}
}
\value{
The replication rate of the result set at 80\% power
}
\description{
This function calculates the replicaiton rate in a test cohort
}
