% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgrm.R
\name{get_summary_stats}
\alias{get_summary_stats}
\title{Retrieve summary statistics}
\usage{
get_summary_stats(dataset = "ukb", exclude_overlap = TRUE)
}
\arguments{
\item{dataset}{A string that indicates the biobank dataset
Options ukb, mgi, bbj, biovu_eur, biovu_afr. Default all}

\item{exclude_overlap}{If TRUE then associations that were originally sourced from the dataset
will be excluded}
}
\value{
A data.table of the summary statistics
}
\description{
This function returns a data.table of summary statics from biobank cohorts
}
\details{
This function retrieves summary statistics from biobank cohorts.
}
