% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map.R
\docType{data}
\name{icdExampleTable}
\alias{icdExampleTable}
\title{Example table of ICD data should be formatted for run_PGRM_assoc()}
\format{
A data.table with the following columns:
\itemize{
\item \code{person_id}: unqiue identifier of subject
\item \code{icd}: Character vector of icd code
\item \code{flag}: Integer of ICD version (either 9 or 10)
\item \code{entry_date}: Date of ICD code
}
}
\usage{
icdExampleTable
}
\description{
Example table of ICD data should be formatted for run_PGRM_assoc()
}
\keyword{datasets}
