% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map.R
\docType{data}
\name{icdPhecodeMap_V1_2}
\alias{icdPhecodeMap_V1_2}
\title{ICD code to phecode map, version 1.2}
\format{
A data.table with the following columns:
\itemize{
\item \code{icd}: Character vector of icd code.
\item \code{flag}: Integer of ICD version (either 9 or 10)
\item \code{phecode}: Character vector of phecode
}
}
\usage{
icdPhecodeMap_V1_2
}
\description{
This map is used in the make_pheno() function
}
\keyword{datasets}
