% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgrm.R
\name{make_GRS}
\alias{make_GRS}
\title{Generate a genetic risk score (GRS) based on variants in the PGRM.}
\usage{
make_GRS(PGRM, geno, phecode, prune = TRUE, R2 = 0.2, LOUD = TRUE)
}
\arguments{
\item{PGRM}{A data.table of the PGRM, generated with get_PGRM()}

\item{geno}{A matrix or 'BEDMatrix' object containing genetic data}

\item{phecode}{A string specifying the phecode}

\item{prune}{If TRUE, then the SNP list is pruded at the R2 threshold specified}

\item{R2}{A numeric value between 0 and 1 indicating the desired R2 value}

\item{LOUD}{If TRUE then progress info is printed to the terminal. Default TRUE}
}
\value{
A data.table with columns \code{person_id} and \code{GRS}
}
\description{
Generate a genetic risk score (GRS) based on variants in the PGRM.
}
