% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgrm.R
\name{make_pheno}
\alias{make_pheno}
\title{Make a phecode phenotype file from table of ICD codes}
\usage{
make_pheno(icds, phecode_version = "V1.2")
}
\arguments{
\item{icds}{A data table ICD codes. Must include columns \code{person_id}, \code{icd}, \code{flag}}

\item{phecode_version}{Phecode map version. Currently only supports "V1.2", which is default}
}
\value{
An numeric value of the percent of associations that are powered
}
\description{
This function calculates \% of powered associations for a test cohort that has been annotated with PGRM.
}
