% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map.R
\docType{data}
\name{PGRM_ALL}
\alias{PGRM_ALL}
\title{The phenotype genotype referene map (PGRM)}
\format{
A data.table with the following columns:
\itemize{
\item \code{assoc_ID}: Integer vector of the unique identifier for each association
\item \code{SNP_hg19}: Character vector variants encoded in hg19 genome reference build
\item \code{SNP_hg38}: Character vector variants encoded in hg38 genome reference build
\item \code{ancestry}: Character vector of the ancestery of the source GWAS using 1000 Genomes superpopulations ancestry groupings
(\strong{AFR}: African, \strong{EAS}: East Asian, \strong{EUR}: European, \strong{AMR}: AdMixed American, \strong{SAS}: South Asian)
\item \code{rsID}: Character indicating the Reference SNP cluster ID
\item \code{risk_allele_dir}: Character indicating the specifying the risk allele
(\strong{alt}: Alternate allele, \strong{ref}: Reference allele)
\item \code{phecode}: Character vector of the phecode for the association
\item \code{phecode_string}: Character vector of the phecode string label
\item \code{cat_LOG10_P}: Numeric vector of the -log10(P) of the association from the catalog
\item \code{cat_OR}: Numeric vector of the odds ratio of the association
\item \code{cat_L95}: Numeric vector of the 95\% lower confidence interval of the association
\item \code{cat_U95}: Numeric vector of the 95\% upper confidence interval of the association
\item \code{Study_accession}: Character vector of the study accession ID from the GWAS catalog
\item \code{cases_needed_AFR}: Integer vector of the estimated cases needed at 80\% power for African ancestry cohort
\item \code{cases_needed_EAS}: Integer vector of the estimated cases needed at 80\% power for East Asian ancestry cohort
\item \code{cases_needed_AMR}: Integer vector of the estimated cases needed at 80\% power for AdMixed American ancestry cohort
\item \code{cases_needed_SAS}: Integer vector of the estimated cases needed at 80\% power for South Asian ancestry cohort
\item \code{cases_needed_ALL}: Integer vector of the estimated cases needed at 80\% power for multi-ancestry cohort
\item \code{category_string}: Character vector of the phecode category
}
}
\usage{
PGRM_ALL
}
\description{
This table provides GWAS associations results from the GWAS catalog, annotated by phecodes, ancestry, etc.
}
\seealso{
\code{\link[=get_PGRM]{get_PGRM()}}
}
\keyword{datasets}
