% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map.R
\docType{data}
\name{phecode_info}
\alias{phecode_info}
\title{Information about phecodes}
\format{
A data table with the following columns:
\itemize{
\item \code{phecode}: Character vector of phecode
\item \code{phecode_string}: Character vector or string label for phecode
\item \code{sex}: Indicates if phecode is sex specific. Values \code{Female}, \code{Male}, or \code{Both}
\item \code{leaf}: A boolean value indicating that the phecode is a leaf (has no children)
\item \code{phecode_top}: Character vector of the parent phecode
\item \code{phecode_top}: Character vector of the category label
}
}
\source{
\url{https://phewascatalog.org}
}
\usage{
phecode_info
}
\description{
The data are artificial and do not correspond to real patients.
}
\seealso{
\code{\link[=get_pheno]{get_pheno()}}
}
\keyword{datasets}
