% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map.R
\docType{data}
\name{results_BBJ}
\alias{results_BBJ}
\title{Summary statistics from BioBank Japan (BBJ)}
\format{
A data.table with the following columns:
\itemize{
\item \code{SNP}: String of SNP identifier, format CHR:position:ref:alt. Build = hg19.
\item \code{phecode}: String of a phecode
\item \code{cases}: Integer of the number of cases for specified phecode
\item \code{controls}: Integer of the number of controls for specified phecode
\item \code{odds_ratio}: Numeric odds ratio
\item \code{P}: Numeric raw P-value
\item \code{L95}: Numeric 95\% lower confidence interval
\item \code{U95}: Numeric 95\% upper confidence interval
\item \code{cohort_match}: Integer that is 1 if the source of the PGRM association includes BBJ, 0 if not
}
}
\source{
\url{https://pheweb.jp}
}
\usage{
results_BBJ
}
\description{
This table provides BBJ GWAS summary statistics for SNPs and phecodes that are in the PGRM
}
\keyword{datasets}
