% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map.R
\docType{data}
\name{results_BioVU_AFR}
\alias{results_BioVU_AFR}
\title{Summary statistics from the African ancestery cohort in BioVU}
\format{
A data.table with the following columns:
\itemize{
\item \code{SNP}: String of SNP identifier, format CHR:position:ref:alt. Build = hg19
\item \code{phecode}: String of a phecode
\item \code{cases}: Integer of the number of cases for specified phecode
\item \code{controls}: Integer of the number of controls for specified phecode
\item \code{odds_ratio}: Numeric odds ratio
\item \code{P}: Numeric raw P-value
\item \code{L95}: Numeric 95\% lower confidence interval
\item \code{U95}: Numeric 95\% upper confidence interval
\item \code{cohort_match}: Integer that is 1 if the source of the PGRM association includes BioVU, 0 if not
}
}
\usage{
results_BioVU_AFR
}
\description{
This table provides BioVU African ancestry GWAS summary statistics for SNPs and phecodes that are in the PGRM
}
\keyword{datasets}
