% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map.R
\docType{data}
\name{results_UKBB}
\alias{results_UKBB}
\title{Summary statistics from UK Biobank (UKBB)}
\format{
A data.table with the following columns:
\itemize{
\item \code{SNP}: String of SNP identifier, format CHR:position:ref:alt. Build = hg38
\item \code{phecode}: String of a phecode
\item \code{cases}: Integer of the number of cases for specified phecode
\item \code{controls}: Integer of the number of controls for specified phecode
\item \code{odds_ratio}: Numeric odds ratio
\item \code{P}: Numeric raw P-value
\item \code{L95}: Numeric 95\% lower confidence interval
\item \code{U95}: Numeric 95\% upper confidence interval
\item \code{cohort_match}: Integer that is 1 if the source of the PGRM association includes UKBB, 0 if not
}
}
\source{
\url{https://pheweb.org/UKB-TOPMed/}
}
\usage{
results_UKBB
}
\description{
This table provides UK Biobank GWAS summary statistics for SNPs and phecodes that are in the PGRM
}
\keyword{datasets}
