% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgrm.R
\name{run_PGRM_assoc}
\alias{run_PGRM_assoc}
\title{Run associations in the pgrm on data from a test cohort. Function requires row level phenotype,
genotype, and covariate information.}
\usage{
run_PGRM_assoc(
  geno,
  pheno,
  demos,
  covariates,
  PGRM,
  MCC = 2,
  minimum_case_count = 100,
  use_exclude_ranges = TRUE,
  check_sex = FALSE,
  LOUD = TRUE
)
}
\arguments{
\item{geno}{A matrix or 'BEDMatrix' object containing genetic data}

\item{pheno}{A data.table having one row per person in the cohort. Must have}

\item{demos}{A data.table having one row per person in the cohort. Must have
a column \code{person_id}. May include other columns used for covariates in later function calls.}

\item{covariates}{A list of covariates to be used in the logistic regression. All covariates
listed must be present in the demos table}

\item{PGRM}{A data.table of the PGRM, generated with get_PGRM()}

\item{MCC}{An integer specifying the minimum code count required for a case (default MCC=2)}

\item{minimum_case_count}{An integer specifiying the minimum number of cases required in the test
cohort to be included in the analysis (default minimum_case_count=100)}

\item{use_exclude_ranges}{If TRUE then exclude ranges are applied to controls}

\item{check_sex}{If TRUE then individuals with sex == F in the demos table will be excluded
from male-specific phenotypes, and vice-versa. Sex specific phecodes are specified in
phecode_info. For this function to work, the demos table must have column \code{sex} and the values
must be 'M' for Male and 'F' for Female}

\item{LOUD}{If TRUE then progress info is printed to the terminal. Default TRUE}
}
\value{
A data.table with annotated results from association tests
}
\description{
(NOTE: If you used principle components as covariates, make sure they are not coded as numeric.
Otherwise the logistic regression will treat them as categorical variables, making the function
run very slowly before failing.)
}
