% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map.R
\docType{data}
\name{summary_stats_anno}
\alias{summary_stats_anno}
\title{Summary statistics from all five test cohorts with annotations}
\format{
A data.table with the following columns:
\itemize{
\item \code{assoc_ID}: Integer vector of the unique identifier for each association
\item \code{phecode}: Character vector of the phecode for the association
\item \code{phecode_string}: Character vector of the phecode string label
\item \code{category_string}: Character vector of the phecode category
\item \code{cases}: Integer of the number of cases for specified phecode/cohort
\item \code{controls}: Integer of the number of controls for specified phecode/cohort
\item \code{odds_ratio}: Numeric odds ratio for test cohort association
\item \code{P}: Numeric raw P-value for test cohort association
\item \code{L95}: Numeric 95\% lower confidence interval for test cohort association
\item \code{U95}: Numeric 95\% upper confidence interval for test cohort association
\item \code{dataset}: Character vector of the source dataset (\strong{BBJ}: Biobank Japan,
\strong{BioVU_AFR}: BioVU, African ancestry, \strong{BioVU_EUR}: BioVU, European ancestry,
\strong{MGI}: Michigan genomics initiative, \strong{UKBB}: UK Biobank)
\item \code{ancestry}: Character vector of the ancestery of the source GWAS using 1000
Genomes superpopulations ancestry groupings (\strong{AFR}: African, \strong{EAS}: East Asian,
\strong{EUR}: European, \strong{AMR}: AdMixed American, \strong{SAS}: South Asian)
\item \code{rsID}: Character indicating the Reference SNP cluster ID
\item \code{risk_allele_dir}: Character indicating the specifying the risk allele direction
(\strong{alt}: Alternate allele, \strong{ref}: Reference allele)
\item \code{RAF}: GnomAD frequency of risk allele, matched to ancestry
\item \code{cat_LOG10_P}: Numeric vector of the -log10(P) of the association from the catalog
\item \code{cat_OR}: Numeric vector of the odds ratio of the association
\item \code{cat_L95}: Numeric vector of the 95\% lower confidence interval of the association
\item \code{cat_U95}: Numeric vector of the 95\% upper confidence interval of the association
\item \code{Study_accession}: Character vector of the study accession ID from the GWAS catalog
\item \code{Powered}: Boolean value indicating if the association is powered at >=80\%
\item \code{rep}: Boolean value indicating if the association replicated
\item \code{CI_overlap}: Character vector indicating the overlap of the confidence intervals from
the catalog and test association
\item \code{pub_count}: The number of times the association has been reported in the catalog
\item \code{first_pub_date}: The first date the association was published
}
}
\usage{
summary_stats_anno
}
\description{
This table summary statistics for PGRM association in BioVU (EUR and AFR), BBJ, MGI, and UKBB.
All associations are annotated and the table is in long format.
}
\keyword{datasets}
