% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgrm.R
\name{annotate_results}
\alias{annotate_results}
\title{Annotate a result set with the PGRM}
\usage{
annotate_results(
  results,
  use_allele_dir = TRUE,
  ancestry = "all",
  build = "hg19",
  phecode_version = "V1.2",
  calculate_power = TRUE,
  annotate_ci_overlap = TRUE,
  LOUD = TRUE
)
}
\arguments{
\item{results}{A data frame (or data.table) with results of a test cohort; columns
for SNP, phecode, cases, controls, odds_ratio, P.}

\item{use_allele_dir}{If TRUE, direction of effect is used when assessing if an
association is replicated. To use this argument, odds ratios must be reported
for the alternative allele}

\item{ancestry}{A string that specifies ancestry of the PGRM that is then used
to annotate the results file.
Options EAS, EUR, AFR, SAS, AMR, ALL. Default ALL}

\item{build}{A string indicating the genome reference build used in the results
table. Options hg19, hg38. Default is hg19.}

\item{phecode_version}{A string indicating the phecode version used in the results
table. (Currently only V1.2 is supported)}

\item{calculate_power}{If TRUE then power calculations will be conducted using
case and control counts from the results file. Necessary for get_AER(). Default FALSE}

\item{annotate_ci_overlap}{If TRUE then a column called 'annotate_ci_overlap'
is added to the table, values:
\strong{overlap}: 95\% CIs of PGRM and test cohort overlap
\strong{test_cohort_greater}: 95\% CI of test cohort greater than PGRM
\strong{PGRM_greater}: 95\% CI of PGRM greater than test cohort
If annotate_ci_overlap is TRUE, then results must include 95\% CIs}

\item{LOUD}{If TRUE then progress info is printed to the terminal. Default FALSE}
}
\value{
A data.table of the results file annotated with columsn from the PGRM
}
\description{
This function annotates a result from a test cohort with information from the PGRM
}
\details{
This function takes a dataframe with summary statistics from a test cohort.
For an example of the way to format the results data frame, look at the fomatting from
\link{summary_stats}. (NOTE: If the direction of effect
is used to determine if an association is replicated, then the odds ratios of
the result set must be oriented to the alternative allele.)

The function returns a data.table with the following annotations:
\itemize{
\item Phecode informtion, including phecode_string and phecode_category
\item Risk allele frequency from GnomAD (column RAF), ancestry specified by the ancestry argument
\item The rsID
\item The direction of effect (ref or alt) and risk allele of the original association
\item Summary statistics from the GWAS catalog association, including the -log10(P),
odds ratio, and 95\% confidence intervals (cat_log10_p, cat_or, cat_l95, cat_u95)
\item The study accession ID from the GWAS catalog
\item A column called powered, which is 1 or 0 indicating whether the test association
is powered > 80\%. If calculate_power==TRUE, then the power is determined by the case/control
counts specified in the results data.table. Otherwise, it is derrived from the estimate
pre-computed cases needed assuming a 1:5 case:control ratio. All power calculations use
alpha=0.0
\item A column called rep that indicates if the association is replicated (i.e.
p<0.05 in the test cohort; if use_allele_dir==TRUE, then the direction of effect
from the test cohort must also be consistant with what is reported in the catalog)
\item If annotate_ci_overlap is true, then information about the relationship
between the 95\% CIs from the catalog and the test set is included in column
ci_overlap, and new columns for odds_ratio, l95, and u95 are created(risk_or, risk_l95,
risk_u95) that are oriented to the risk allele. (This option requires that the confidence
intervals are reported in the test cohort summary statistics)
}
}
\examples{
library(pgrm)

## get summary statistics for UK Biobanks
summary_stats_ukb = get_summary_stats(dataset = 'ukb')

## Filter out associations where UKB was included in the original study
summary_stats_ukb = summary_stats_ukb[cohort_match == 0]

## annotate summary statistics with pgrm
anno = annotate_results(head(summary_stats_ukb,n=10), ancestry = 'EUR',
  build = 'hg38', calculate_power = TRUE)

## Get the replication rate of associations powered at >80\%
get_RR(anno)

## Get the replication rate of all associations
get_RR(anno,include='all')

## Get the actual:expected ratio
get_AER(anno)
}
\seealso{
Example result sets: \link{summary_stats}
}
