% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgrm.R
\name{get_PGRM}
\alias{get_PGRM}
\title{Get instance of PGRM}
\usage{
get_PGRM(ancestry = "all", build = "hg19", phecode_version = "V1.2")
}
\arguments{
\item{ancestry}{A string that indicates the ancestry of the PGRM. Options EAS, EUR, AFR, SAS, AMR, ALL. Default ALL}

\item{build}{A string indicating the genome reference build. Options hg19, hg37. Default is hg19.}

\item{phecode_version}{A string indicating the phecode version. Currently only V1.2 is supported, which is the default}
}
\value{
A data.table of the PGRM
}
\description{
This function generates a PGRM copy with specified ancestry, build, and phecode version
}
