% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgrm.R
\name{get_pheno}
\alias{get_pheno}
\title{Add case/control status for a specified phenotype to a covariate file.}
\usage{
get_pheno(
  pheno,
  covars,
  phecode,
  MCC = 2,
  use_exclude_ranges = TRUE,
  check_sex = FALSE
)
}
\arguments{
\item{pheno}{A data.table with phecode phenotypes. columns \code{person_id}, \code{phecode}, and \code{N}
required. The phecode columns should be a character vector. Column \code{N} specifies the number of
unique dates a phecode occurs for that individual. Only phecodes that are present for an individual
are included (i.e. N is always greater than one).}

\item{covars}{A data.table having one row per person in the cohort. Must havea column names
\code{person_id}. Must include other columns specified in covariate_names.}

\item{phecode}{A character vector of the phecode}

\item{MCC}{A numeric value specified the minimum code count required for cases.}

\item{use_exclude_ranges}{If TRUE then individuals who are not cases but have a phecode within
the exclude ranges are excluded.}

\item{check_sex}{If TRUE then the \code{pheno} column is set to zero for individuals with a sex that
does not match the specified sex. This argument is only relevant for sex specific codes
(e.g. Prostate cancer). To use this function, sex must be included as a column in the covars data.table.}
}
\value{
The covar data.table with a new column called \code{pheno}. This column is 0 if the individual,
is a control for specified phecode, 1 if they are a case, and NA if they are excluded.
}
\description{
Add case/control status for a specified phenotype to a covariate file.
}
