% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgrm.R
\name{get_RR}
\alias{get_RR}
\title{Calculate the replicaiton rate (RR) of a test cohort}
\usage{
get_RR(annotated_results, include = "powered", LOUD = TRUE)
}
\arguments{
\item{annotated_results}{A data table of results that have been annotated with the PGRM}

\item{include}{A character string. If "powered" then only powered associations are included (default). If "all" then all associations are included}

\item{LOUD}{If TRUE then progress info is printed to the terminal. Default TRUE}
}
\value{
An numeric value of the replication rate of the result set
}
\description{
This function calculates the replicaiton rate in a test cohort that has been annotated with PGRM. By default, it calculates the RR for associations that powered at >80\%.
}
