% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgrm.R
\name{get_RR}
\alias{get_RR}
\title{Calculate the replicaiton rate (RR) of a test cohort}
\usage{
get_RR(annotated_results, include = "powered", LOUD = TRUE)
}
\arguments{
\item{annotated_results}{A data table of results that have been annotated with the PGRM}

\item{include}{A character string. If "powered" then only powered associations are included.
If "all" then all associations are included}

\item{LOUD}{If TRUE then progress info is printed to the terminal. Default TRUE}
}
\value{
An numeric value of the replication rate of the result set
}
\description{
This function calculates the replicaiton rate in a test cohort that has been annotated with PGRM.
By default, it calculates the RR for associations that powered at >80\%.
}
\examples{
  library(pgrm)
  library(data.table)

  ## Read in an ICD file. Format should look like the example file.
  ## Note: Don't forget to load icd column as a character vector!
  head(icdExampleTable)

  ## make pheno table
  pheno=make_pheno(icdExampleTable)
  head(pheno)

  ## Load in covariate file. Here's a generated example
  covar_table_test = data.table(
  person_id = 1:4, sex = c('F', 'M', 'F', 'M'),
  last_age = c(28772,18028,11636,14589))

  ## Load genotype file
  library(gaston)
 # geno=read.bed.matrix('data/geno_test')

  ## Make a PGRM instance, specifying genome build and ancestry
  PGRM=get_PGRM(build='hg19',ancestry='all')

  ## Run associations from the PGRM
 # run_PGRM_assoc(geno,pheno,covars,covariate_names=c('last_age'),PGRM,MCC=2, use_exclude_range=TRUE,check_sex=TRUE)
  
}
