% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map.R
\docType{data}
\name{PGRM_ALL}
\alias{PGRM_ALL}
\title{The phenotype genotype referene map (PGRM)}
\format{
A data.table with the following columns:
\itemize{
\item \code{assoc_ID}: Integer vector of the unique identifier for each association
\item \code{SNP_hg19}: Character vector variants encoded in hg19 genome reference build
\item \code{SNP_hg38}: Character vector variants encoded in hg38 genome reference build
\item \code{rsID}: Character indicating the Reference SNP cluster ID
\item \code{risk_allele_dir}: Character vector specifying the risk allele direction
(\strong{alt}: Alternate allele, \strong{ref}: Reference allele)
\item \code{risk_allele}: Character specifying the risk allele
\item \code{phecode}: Character vector of the phecode for the association
\item \code{phecode_string}: Character vector of the phecode string label
\item \code{category_string}: Character vector of the phecode category
\item \code{ancestry}: Character vector of the ancestery of the source GWAS using 1000
Genomes superpopulations ancestry groupings (\strong{AFR}: African, \strong{EAS}: East Asian,
\strong{EUR}: European, \strong{AMR}: AdMixed American, \strong{SAS}: South Asian)
\item \code{cat_LOG10_P}: Numeric vector of the -log10(P) of the association from the catalog
\item \code{cat_OR}: Numeric value of the odds ratio of the association
\item \code{cat_L95}: Numeric value of the 95\% lower confidence interval of the association
\item \code{cat_U95}: Numeric value of the 95\% upper confidence interval of the association
\item \code{Study_accession}: Character vector of the study accession ID from the GWAS catalog
\item \code{pubmedid}: Numeric value of the PubMed ID for source publication of association
\item \code{pub_count}: Numeric value indicating how many times association has been published in GWAS catalog
\item \code{first_pub_date}: Date of the first date the association was published
\item \code{AFR_RAF}: Numeric value of the African ancestry risk allele frequency from gnomAD
\item \code{EUR_RAF}: Numeric value of the European (non-Finnish) ancestry risk allele frequency from gnomAD
\item \code{EAS_RAF}: Numeric value of the East Asian ancestry risk allele frequency from gnomAD
\item \code{AMR_RAF}: Numeric value of the Latino/Admixed American ancestry risk allele frequency from gnomAD
\item \code{SAS_RAF}: Numeric value of the South Asian ancestry risk allele frequency from gnomAD
\item \code{ALL_RAF}: Numeric value of the risk allele frequency from gnomAD
}
}
\source{
Allele frequencies: \url{https://gnomad.broadinstitute.org/}
Summary statistics: \url{https://www.ebi.ac.uk/gwas/}
}
\usage{
PGRM_ALL
}
\description{
This table provides GWAS associations results from the GWAS catalog, annotated by phecodes, ancestry, etc.
}
\details{
The odds ratio and 95\% confidence intervals (columns cat_OR, cat_L95,
and cat_U95) are reported relative to the risk allele which is specified in the
column risk_allele.
}
\seealso{
\code{\link[=get_PGRM]{get_PGRM()}}
}
\keyword{datasets}
