% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgrm.R
\name{run_PGRM_assoc}
\alias{run_PGRM_assoc}
\title{Run association tests for phenotype/genotype pairs in the PGRM}
\usage{
run_PGRM_assoc(
  genotypes,
  pheno,
  demos,
  covariates,
  PGRM,
  MCC = 2,
  minimum_case_count = 100,
  LOUD = TRUE,
  use_exclude_ranges = F
)
}
\arguments{
\item{genotypes}{A 'BEDMatrix' object linked to the PLINK bed file containing genetic data. The row names correspond to the person_id's in demos and scores tables. The column names correspond to variant IDs.}

\item{pheno}{A data.table of phenotypes to be used in the association analysis. Must have columns person_id, phecodes}

\item{demos}{A data.table of covariates to be used in the association analysis. Must have column person_id and columns for every covariate in covariates list.}

\item{PGRM}{A copy of the PGRM. Can be generated with function get_PGRM()}

\item{MCC}{The minimum code count needed for cases. Default == 2.}

\item{minimum_case_count}{The minimum number of cases required for an association test to be conducted.}

\item{LOUD}{If TRUE then progress info is printed to the terminal. Default FALSE}

\item{use_exclude_range}{If TRUE then exclude ranges are applied to controls.}
}
\value{
A data.table of association results for all eligable phenotype/genotype pairs in the PGRM.
}
\description{
This function takes row level data from a test cohort and runs association tests for all phenotype/genotype pairs included in the PGRM
}
\details{
This function takes row level data from a test cohort as well as a copy of the PGRM. It runs an association analysis for each phenotype/genotype pair in the PGRM
and returns the results.
}
\examples{
library(pgrm)
## Get a copy of the PGRM for build hg19, East Asian ancestry
PGRM_EAS=get_PGRM(build="hg19",ancestry="EAS")

## Read in a phenotype file
pheno=read.csv(pheno_file,header=TRUE,colClasses=c("character","character","integer"),stringsAsFactors = F)

## Read in demographics file
demos=read.csv(demos_file,header=TRUE,stringsAsFactors = F)

## Read in genotype file
genotypes=read.bed.matrix(geno_file)
## Filter out genotypes with low callrate
genotypes=select.snps(geno, callrate > 0.98)

covariates=c('sex','last_age','PC1','PC2','PC3','PC4','PC5','PC6','PC7','PC8')
results_MCC2 = run_PGRM_assoc(geno=genotypes, pheno=pheno, demos=demos, covariates=covariates,PGRM=PGRM_EAS,MCC=2,minimum_case_count=100,use_exclude_range=T,LOUD=T)

}
